+++ tests/browser/test_markets.py
@@ -12,12 +12,15 @@ from tests.browser.common_selectors import (
 )
 from tests.browser.util import (
     attach_jpg_screenshot,
-    should_not_see,
+    should_not_see_any_element,
     should_not_see_errors,
     should_see_all_elements,
 )
 
-pytestmark = pytest.mark.browser
+pytestmark = [
+    pytest.mark.browser,
+    pytest.mark.markets,
+]
 
 
 @pytest.mark.django_db
@@ -41,6 +44,6 @@ def test_can_view_markets_as_signed_in_user(
     browser.get(live_server.url + '/markets/')
 
     attach_jpg_screenshot(browser, 'Markets')
-    should_not_see(browser, DashboardModalLetsGetToKnowYou)
+    should_not_see_any_element(browser, DashboardModalLetsGetToKnowYou)
     should_see_all_elements(browser, HeaderSignedIn)
     should_see_all_elements(browser, MarketsContainer)