+++ tests/conftest.py
@@ -11,17 +11,25 @@ from wagtail_factories import PageFactory, SiteFactory
 
 import tests.unit.domestic.factories
 import tests.unit.exportplan.factories
-from core import helpers
+from core import helpers as core_helpers
+from core.management.commands.create_tours import defaults as tour_steps
+from core.models import Tour
 from directory_api_client import api_client
+from directory_constants import choices
+from exportplan import helpers as exportplan_helpers
+from sso import helpers as sso_helpers
 from sso.models import BusinessSSOUser
 from tests.helpers import create_response
+from tests.unit.learn import factories as learn_factories
 
 # This is to reduce logging verbosity of these two libraries when running pytests
 # with DEBUG=true and --log-cli-level=DEBUG
 selenium_logger = logging.getLogger('selenium')
 pil_logger = logging.getLogger('PIL')
+urllib3_logger = logging.getLogger('urllib3')
 selenium_logger.setLevel(logging.CRITICAL)
 pil_logger.setLevel(logging.CRITICAL)
+urllib3_logger.setLevel(logging.CRITICAL)
 
 
 @pytest.fixture
@@ -57,10 +65,18 @@ def domestic_site(domestic_homepage, client):
 
 
 @pytest.fixture
-def domestic_site_browser_tests(domestic_homepage, exportplan_dashboard, client):
+def domestic_site_browser_tests(live_server, domestic_homepage, exportplan_dashboard, client):
+    """Will server domestic site on the same port as liver_server.
+    Note:
+        live_server.url looks like this: http://localhost:48049
+        The value of live_server.url can be also set via --liveserver parameter:
+        make ARGUMENTS="--liveserver=localhost:48049'" pytest_browser
+    """
+    live_server_port = int(live_server.url.split(':')[-1])
     return SiteFactory(
         root_page=domestic_homepage,
         hostname='localhost',  # This allows Browser to access site via live_server.url
+        port=live_server_port  # This forces Site to be server on the same port as live_server
     )
 
 
@@ -112,20 +128,20 @@ def mock_airtable_rules_regs():
             'id': '1',
             'fields':
                 {
-                    'Country': 'India',
-                    'Export Duty': 1.5,
-                    'Commodity code': '2208.50.12',
-                    'Commodity Name': 'Gin and Geneva 2l'
+                    'country': 'India',
+                    'export_duty': 1.5,
+                    'commodity_code': '2208.50.12',
+                    'commodity_name': 'Gin and Geneva 2l'
                 },
         },
         {
             'id': '2',
             'fields':
                 {
-                    'Country': 'China',
-                    'Export Duty': 1.5,
-                    'Commodity code': '2208.50.13',
-                    'Commodity Name': 'Gin and Geneva'
+                    'country': 'China',
+                    'export_duty': 1.5,
+                    'commodity_code': '2208.50.13',
+                    'commodity_name': 'Gin and Geneva'
                 },
         },
     ]
@@ -140,6 +156,113 @@ def mock_user_location_create():
     stub = mock.patch.object(api_client.personalisation, 'user_location_create', return_value=response)
     yield stub.start()
     stub.stop()
+
+
+@pytest.fixture
+@pytest.mark.django_db(transaction=True)
+@mock.patch.object(core_helpers, 'get_dashboard_export_opportunities')
+@mock.patch.object(core_helpers, 'get_dashboard_events')
+@mock.patch.object(sso_helpers, 'get_company_profile')
+@mock.patch.object(core_helpers, 'get_markets_page_title')
+def mock_dashboard_profile_events_opportunities(
+        mock_get_markets_page_title,
+        mock_get_company_profile,
+        mock_get_dashboard_events,
+        mock_get_dashboard_export_opportunities,
+):
+    mock_get_markets_page_title.return_value = 'Some page title'
+    mock_get_company_profile.return_value = {
+        'expertise_countries': ['AF'], 'expertise_industries': [choices.SECTORS[0][0]]
+    }
+    mock_get_dashboard_events.return_value = []
+    mock_get_dashboard_export_opportunities.return_value = []
+
+
+@pytest.fixture
+@pytest.mark.django_db(transaction=True)
+@mock.patch.object(exportplan_helpers, 'get_exportplan_marketdata')
+@mock.patch.object(api_client.dataservices, 'get_lastyearimportdata')
+@mock.patch.object(api_client.dataservices, 'get_corruption_perceptions_index')
+@mock.patch.object(api_client.dataservices, 'get_easeofdoingbusiness')
+@mock.patch.object(api_client.exportplan, 'exportplan_list')
+def mock_export_plan_requests(
+        mock_export_plan_list,
+        mock_ease_of_doing_business,
+        mock_get_corruption_perceptions_index,
+        mock_get_last_year_import_data,
+        mock_get_export_plan_market_data,
+):
+    data = [
+        {
+            'export_countries': ['UK'],
+            'export_commodity_codes': [100],
+            'rules_regulations': {'rule1': 'AAA'}
+        }
+    ]
+    mock_export_plan_list.return_value = create_response(data)
+
+    ease_of_doing_business_data = {
+        'country_name': 'China',
+        'country_code': 'CHN',
+        'cpi_score_2019': 41,
+        'rank': 80,
+    }
+    mock_ease_of_doing_business.return_value = create_response(
+        status_code=200,
+        json_body=ease_of_doing_business_data,
+    )
+
+    cpi_data = {
+        'country_name': 'China',
+        'country_code': 'CHN',
+        'cpi_score_2019': 41,
+        'rank': 80,
+    }
+    mock_get_corruption_perceptions_index.return_value = create_response(status_code=200, json_body=cpi_data)
+
+    mock_get_last_year_import_data.return_value = create_response(
+        status_code=200, json_body={'lastyear_history': 123}
+    )
+
+    mock_get_export_plan_market_data.return_value = {'timezone': 'Asia/Shanghai', }
+
+
+@pytest.mark.django_db(transaction=True)
+@pytest.fixture
+def topics_with_lessons(domestic_site_browser_tests):
+    domestic_homepage = domestic_site_browser_tests.root_page
+    topic_a = learn_factories.TopicPageFactory(
+        parent=domestic_homepage, title='Lesson topic A', slug='topic-a',
+    )
+    lesson_a1 = learn_factories.LessonPageFactory(
+        parent=topic_a, title='Lesson A1', slug='lesson-a1',
+    )
+    lesson_a2 = learn_factories.LessonPageFactory(
+        parent=topic_a, title='Lesson A2', slug='lesson-a2',
+    )
+
+    topic_b = learn_factories.TopicPageFactory(
+        parent=domestic_homepage, title='Lesson topic B', slug='topic-b',
+    )
+    lesson_b1 = learn_factories.LessonPageFactory(
+        parent=topic_b, title='Lesson B1', slug='lesson-b1',
+    )
+    lesson_b2 = learn_factories.LessonPageFactory(
+        parent=topic_b, title='Lesson B2', slug='lesson-b2',
+    )
+    return [(topic_a, [lesson_a1, lesson_a2]), (topic_b, [lesson_b1, lesson_b2])]
+
+
+@pytest.fixture
+def mock_export_plan_dashboard_page_tours(exportplan_dashboard):
+    """Create Export Plan Dashboard page tour steps in reversed order.
+
+    For some reason when page tour steps are created during a unit test run then
+    those steps are shown in reversed order. So in order to show them in the right
+    order they have to be reverse here.
+    """
+    tour_steps.update({'steps': list(reversed(tour_steps['steps']))})
+    return Tour.objects.get_or_create(page=exportplan_dashboard, defaults=tour_steps)
 
 
 @pytest.fixture(scope='session')
@@ -190,6 +313,36 @@ def mock_get_company_profile(patch_get_company_profile):
 
 
 @pytest.fixture
+def patch_get_dashboard_events():
+    yield mock.patch('core.helpers.get_dashboard_events', return_value=None)
+
+
+@pytest.fixture(autouse=True)
+def mock_get_events(patch_get_dashboard_events):
+    yield patch_get_dashboard_events.start()
+    try:
+        patch_get_dashboard_events.stop()
+    except RuntimeError:
+        # may already be stopped explicitly in a test
+        pass
+
+
+@pytest.fixture
+def patch_get_dashboard_export_opportunities():
+    yield mock.patch('core.helpers.get_dashboard_export_opportunities', return_value=None)
+
+
+@pytest.fixture(autouse=True)
+def mock_get_export_opportunities(patch_get_dashboard_export_opportunities):
+    yield patch_get_dashboard_export_opportunities.start()
+    try:
+        patch_get_dashboard_export_opportunities.stop()
+    except RuntimeError:
+        # may already be stopped explicitly in a test
+        pass
+
+
+@pytest.fixture
 def server_user_browser(browser, live_server, user, client):
     client.force_login(user)
     return live_server, user, browser
@@ -219,17 +372,10 @@ def single_opportunity():
 
 
 @pytest.fixture
-@mock.patch.object(helpers, 'get_dashboard_export_opportunities')
-@mock.patch.object(helpers, 'get_dashboard_events')
 def server_user_browser_dashboard(
-    mock_get_dashboard_events, mock_get_dashboard_export_opportunities,
     mock_get_company_profile, server_user_browser, settings, domestic_site_browser_tests
 ):
     live_server, user, browser = server_user_browser
-
-    # these mocked endpoints are occasionally called before user is even logged in
-    mock_get_dashboard_events.return_value = []
-    mock_get_dashboard_export_opportunities.return_value = []
 
     browser.get(f'{live_server.url}/dashboard/')