+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Pre-release
 
+## Hotfix
+- No ticket - v3 ci pipeline manifest.yml file fix
+
 ### Fixed bugs
 
 ### Implemented enhancements
@@ -44,3 +47,7 @@
 - MVP-215 - Track lesson read progress
 - MVP-332 - Content segmentation on sector of interest
 - MVP-349 - Adds capabilities landing page under /learn
+- MVP-328 - Sector modal iteration
+- MVP-329 - Country modal iteration
+- MVP-344 - Wagtail cache
+- MVP-356 - marketdata target-markets